restore non-react-conformance implementation of as prop tests
